delete history mixin
